// src/utils/util-types.ts
/* -------------------------------------------------------------------------- */
/*                                 Extraction                                 */
/* -------------------------------------------------------------------------- */

export type Spectrum1D = {
	wavelength: number;
	flux: number;
	error: number;
	fluxPlusErr: number;
	fluxMinusErr: number;
};

/* -------------------------------------------------------------------------- */
/*                                    Plot                                    */
/* -------------------------------------------------------------------------- */

/* -------------------------------------------------------------------------- */
/*                                 Projection                                 */
/* -------------------------------------------------------------------------- */

export type Projected = {
	x: number;
	y: number;
	z: number;
	visible: boolean;
};

/* -------------------------------------------------------------------------- */
/*                                 Wavelength                                 */
/* -------------------------------------------------------------------------- */